rag/requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.2
numpy>=1.21.0
flask>=2.0.0
tqdm>=4.60.0
datasets>=2.0.0
torch>=1.9.0
requests